feat: propagate the max-cost calculation logic to all tools
Propagates the logic introduced in `prediction_request_lite` (#58) to the rest of the tools.

+++ packages/valory/customs/prediction_request_lite/prediction_request.py
@@ -47,7 +47,7 @@ from tiktoken import encoding_for_model
 
 MechResponseWithKeys = Tuple[str, Optional[str], Optional[Dict[str, Any]], Any, Any]
 MechResponse = Tuple[str, Optional[str], Optional[Dict[str, Any]], Any]
-MaxCostResponse = int
+MaxCostResponse = float
 
 
 TOKEN_COSTS_PER_MODEL_ATTR = "TOKEN_PRICES"
